adding dt to data, and moving methods to test util class

+++ condition-checker/src/main/java/com/mersey/rowing/club/condition_checker/model/openweatherapi/WeatherData.java
@@ -15,7 +15,7 @@ import java.util.List;
 public class WeatherData {
 
     @JsonProperty("dt")
-    private double epochDateTime;
+    private int epochDateTime;
 
     @JsonProperty("temp")
     private double temperature;
